refactor: hoist split_vec_min_alloc into datafusion-common
Per review on #22416: move the split_vec_min_alloc helper out of
functions-aggregate's min_max.rs into datafusion-common's utils
module, so it lives in one place and can be shared.

The function and its five tests move verbatim. min_max_struct.rs and
min_max_bytes.rs now import it from datafusion_common::utils instead
of from the local module.
